+++ tests/proc-macro/tests/loopscale.rs
@@ -25,3 +25,18 @@ fn instruction_dispatch_index_is_some() {
         "expected InstructionDispatch message index for a program with instructions"
     );
 }
+
+#[test]
+fn check_json_serialization() {
+    // account
+    let asset = loopscale::AssetData::default();
+    let json_str = serde_json::to_string(&asset).expect("failed to json serialize");
+    let _: loopscale::AssetData =
+        serde_json::from_str(&json_str).expect("failed to json deserialize");
+
+    // instruction
+    let params = loopscale::CreateStrategyParams::default();
+    let json_str = serde_json::to_string(&params).expect("failed to json serialize");
+    let _: loopscale::CreateStrategyParams =
+        serde_json::from_str(&json_str).expect("failed to json deserialize");
+}